Handle crash when a private index doesn't support JSON
We use Elastic PyPI as internal private index and it seems it
doesn't support the JSON API yet:
https://github.com/khornberg/elasticpypi/issues/48

Therefore it returns a 200 response but not formatted as PyUP expects
it's plain HTML rather than being a valid JSON. In this cas, the whole
update crashes. This is changing the behaviour to ignore the package
in the private index and proceed with the rest of the requirements.