Regex case-insensitive.
Recently we have hit a problem on https://github.com/mozilla-releng/build-puppet when we asked PyUp to make PRs "every week on Wednesday". 
The capitalization was a problem, which in return stopped PyUp to do what we expected it to do.

All I did was to add ". re.IGNORECASE" (I couldhave used re.I - but I find it much cleaner to use the long name, as people will understand what's happening.)