Rework how we do compatible releases
Currently, compatible releases - that is, lines like 'Django~=1.7.2' -
are treat like fixed version - 'Django==1.7.2'. This is incorrect
behavior. In reality, 'Django~1.7.2' resolves to 'Django>=1.7.2,<1.8.0'
and these version specifiers can be used to avoid having to bump a
version repeatedly by ensuring the latest PATCH release of a
semantically versioned package is always installed.

Change how we treat these releases so that compatible packages behave as
expected. We do this by replacing the '~=A' format with '>=A,<B' format
specifiers, allowing us to use all the logic we have built up for
range-based specifiers.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Closes: #361