Commit messages can now be configured via config.
The default values for commit messages stayed the same, but are
now always read from the config (instead of magic strings in the
`Update.get_commit_message` method).

This introduces two new settings for the config:
- `commit_message_template_pin` (for pinning commits)
- `commit_message_template_update` (for update commits)

They should be set to strings. Within these strings the substrings
"{package_name}", "{old_version}" and "{new_version}" will be
replaced with their respective values.

### Technical

The unittests are testing variants of the default messages. This
ensures that the config values are really read from the config
instead of assuming the default values.

The method Update.get_commit_message is not a classmethod anymore. This is because there can't be commit messages without providing a config from which the template can be read. This seemed like the simplest solution to me.