Add a cli flag for running as an integration
Currently there is no way to enable the `integration` flag which the
providers uses to filter out some actions.

It is possible to run the bot using an oauth app e.g. via github
actions, however it will fail due to having no access to endpoints such
as /user/emails.

With this PR supplying the `--integration` will enable restricted tokens
to function as expected.